% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.r
\name{frag_stat}
\alias{frag_stat}
\title{Compute quantiles, mean and standard deviation for two electrodes group marked as soz non marked as soz}
\usage{
frag_stat(frag, elecsoz)
}
\arguments{
\item{frag}{Numeric. Fragility matrix. Row Electrode names. Column index}

\item{elecsoz}{Integer.  Vector soz electrodes (for good electrodes)}

\item{t_step}{Integer. Fragility step size}

\item{fs}{Numeric. Acquisition frequency (Hz)}

\item{time_window_ictal}{Numeric.Time window around seizure onset (s)}
}
\value{
quantile matrix, mean and sdv from both electrodes groups
}
\description{
Compute quantiles, mean and standard deviation for two electrodes group marked as soz non marked as soz
}
\examples{
data("fragm3sp5s")
data("elecsoz")
}
