% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_frag_distribution}
\alias{plot_frag_distribution}
\title{Plot Fragility time distribution for two electrodes group marked and non-marked as soz}
\usage{
plot_frag_distribution(
  cmeansoz,
  cmeansozc,
  csdsoz,
  csdsozc,
  time_window_ictal,
  subject_code,
  j
)
}
\arguments{
\item{cmeansoz}{time mean of soz group}

\item{cmeansozc}{time mean of non soz group}

\item{csdsoz}{time standard deviation of soz group}

\item{csdsozc}{time standard deviation of non soz group}

\item{time_window_ictal}{time window of processed ictal iEEG}

\item{subject_code}{subject code}

\item{j}{seizure number}
}
\value{
plot fragility distribution
}
\description{
Plot Fragility time distribution for two electrodes group marked and non-marked as soz
}
\examples{
data("fragstat")
time_window_ictal=c(-10,10)
plot_frag_distribution( cmeansoz=fragstat[[2]],cmeansozc=fragstat[[3]],csdsoz=fragstat[[4]],csdsozc=fragstat[[5]],time_window_ictal=time_window_ictal,subject_code='PT01',j=1)
}
