% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_frag_quantile}
\alias{plot_frag_quantile}
\title{Plot Fragility time quantiles for two electrodes group marked as soz non marked as soz}
\usage{
plot_frag_quantile(quantilematrixsozsozc, time_window_ictal, subject_code, j)
}
\arguments{
\item{quantilematrixsozsozc}{quantile matrix for the two groupd}

\item{time_window_ictal}{time window of processed ictal iEEG}

\item{subject_code}{subject code}

\item{j}{seizure number}
}
\value{
Quantile plot
}
\description{
Plot Fragility time quantiles for two electrodes group marked as soz non marked as soz
}
\examples{
data("fragstat")
time_window_ictal=c(-10,10)
plot_frag_quantile( quantilematrixsozsozc=fragstat[[1]], time_window_ictal=time_window_ictal,subject_code='pt01',j=1)
}
